moving handler to controller
