+++ src/controllers/api.controller.js
@@ -1,21 +1,19 @@
-import artistsService from "../services/artists.service.js";
+import trackService from "../services/tracks.service.js";
 import albumService from "../services/albums.service.js";
+import artistService from "../services/artists.service.js";
 
 export default {
-	async getAlbumInfo(req, res) {
-		const trackTitle = req.params.title;
-		// console.log("/api/album/:title >>>", trackTitle);
-		try {
-			const album = await albumService.findByTrackTitle(trackTitle);
-			const track = await artistsService.findByTrackTitle(trackTitle);
-			if (!album.album_id) {
-				res.json(track);
-				return;
-			}
-			res.json(album); // Send album data as JSON
-		} catch (error) {
-			console.error(error);
-			res.status(500).send("Internal Server Error");
-		}
+	async getTrackInfo(req, res) {
+		const track_id = req.params.id;
+
+		const track = await trackService.findById(track_id);
+		const album = await albumService.findByTrackId(track_id);
+		const artist = await artistService.findByTrackId(track_id);
+
+		const ret = { ...track, ...album, ...artist };
+
+		console.log("RET: ", ret);
+
+		res.status(200).json(ret);
 	},
 };
